requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
PyJWT==2.4.0
PyMySQL==1.0.2
pytz==2022.2.1
sqlparse==0.4.2
